+++ src/main/java/org/launchcode/demo/controllers/BookswapEmailController.java
@@ -5,6 +5,7 @@ import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
 import org.launchcode.demo.data.UserRepository;
 import org.launchcode.demo.models.User;
+import org.launchcode.demo.models.dto.UserEmailDTO;
 import org.launchcode.demo.models.dto.VerificationFormDTO;
 import org.launchcode.demo.models.email.EmailDetails;
 import org.launchcode.demo.models.email.EmailService;
@@ -15,6 +16,7 @@ import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 
 import java.util.Optional;
 
@@ -55,31 +57,24 @@ public class BookswapEmailController {
             return status;
         }
 
-    @GetMapping("user/verification")
-    public String displayVerificationForm(Model model) {
-        model.addAttribute(new VerificationFormDTO());
-        return "user/verification";
+    @GetMapping("/user/email")
+    public String displayUserEmailForm(Model model){
+        model.addAttribute(new UserEmailDTO());
+        return "user/email";
     }
-    @PostMapping("user/verification")
-    public String processVerificationForm(@ModelAttribute @Valid VerificationFormDTO verificationFormDTO,
-                                          Errors errors, HttpServletRequest request,
-                                          Model model) {
-        String username = verificationFormDTO.getUsername();
-        User user = userRepository.findByUsername(username);
-        String submittedCode = verificationFormDTO.getVerifyCode();
-        String sentCode = user.getVerificationCode();
-        if (!submittedCode.equals(sentCode)) {
-            ///errors.rejectValue("code", "incorrect code", "Codes do not match");
-            model.addAttribute("incorrect", "Incorrect Verification Code. Please Check Email");
-            return "user/verification";
-        }
 
-        user.setIsVerified(true);
-        userRepository.save(user);
+    @PostMapping("/user/email")
+    public String processUserEmailForm(@ModelAttribute @Valid UserEmailDTO userEmailDTO, @RequestParam(value="book") String book, @RequestParam(value="recipient") String recipient,
+                                       Errors errors, HttpServletRequest request,
+                                       Model model) {
 
-        model.addAttribute("username", user.getUsername());
+        String email = userEmailDTO.getEmail();
 
-        return "user/index";
+        EmailDetails theseDetails = new EmailDetails(recipient, "Hello! A user is interested in " + book + ". To confirm swap details, please email " + email, "Someone Is Interested in Your Book!");
+        sendUserMail(theseDetails);
 
+        model.addAttribute("success", "Email sent!");
+        return "user/email";
     }
+
 }
